validation related bug fixes
